+++ internal/cmd/install_mailhog.go
@@ -11,6 +11,7 @@ import (
 	"github.com/craftcms/nitro/internal/scripts"
 )
 
+// TODO move this to the nitrod API
 var mailhogCommand = &cobra.Command{
 	Use:   "mailhog",
 	Short: "Install mailhog",