+++ internal/cmd/root.go
@@ -47,7 +47,11 @@ func init() {
 		dbCommand,
 		completionCmd,
 		installCommand,
+		phpCommand,
+		nginxCommand,
 	)
+	phpCommand.AddCommand(phpRestartCommand, phpStartCommand, phpStopCommand, inisetCommand)
+	nginxCommand.AddCommand(nginxStartCommand, nginxStopCommand, nginxRestartCommand)
 	xdebugCommand.AddCommand(xdebugOnCommand, xdebugOffCommand)
 }