+++ internal/task/apply_test.go
@@ -118,6 +118,31 @@ func TestApply(t *testing.T) {
 					UseSyscall: false,
 					Args:       []string{"exec", "mytestmachine", "--", "sudo", "apt-get", "install", "-y", "php7.4", "php7.4-mbstring", "php7.4-cli", "php7.4-curl", "php7.4-fpm", "php7.4-gd", "php7.4-intl", "php7.4-json", "php7.4-mysql", "php7.4-pgsql", "php7.4-zip", "php7.4-xml", "php7.4-soap", "php7.4-bcmath", "php7.4-gmp", "php-xdebug", "php-imagick", "blackfire-agent", "blackfire-php"},
 				},
+				{
+					Type:       "exec",
+					UseSyscall: false,
+					Args:       []string{"exec", "mytestmachine", "--", "sudo", "update-alternatives", "--set", "php", "/usr/bin/php7.4"},
+				},
+				{
+					Type:       "exec",
+					UseSyscall: false,
+					Args:       []string{"exec", "mytestmachine", "--", "sudo", "update-alternatives", "--set", "phar", "/usr/bin/phar7.4"},
+				},
+				{
+					Type:       "exec",
+					UseSyscall: false,
+					Args:       []string{"exec", "mytestmachine", "--", "sudo", "update-alternatives", "--set", "phar.phar", "/usr/bin/phar.phar7.4"},
+				},
+				{
+					Type:       "exec",
+					UseSyscall: false,
+					Args:       []string{"exec", "mytestmachine", "--", "sudo", "update-alternatives", "--set", "phpize", "/usr/bin/phpize7.4"},
+				},
+				{
+					Type:       "exec",
+					UseSyscall: false,
+					Args:       []string{"exec", "mytestmachine", "--", "sudo", "update-alternatives", "--set", "php-config", "/usr/bin/php-config7.4"},
+				},
 			},
 		},
 		{